chore: organize folder structure
