+++ internal/domain/subscription/errors.go
@@ -1,4 +1,4 @@
-package order
+package subscription
 
 import (
 	"net/http"
