+++ internal/domain/subscription/subscriptionHandler.go
@@ -1,0 +1,1 @@
+package subscription
