+++ internal/domain/subscription/subscriptionUsecase.go
@@ -1,0 +1,1 @@
+package subscription
