+++ internal/domain/summary/summaryHandler.go
@@ -1,0 +1,1 @@
+package summary
