+++ internal/domain/summary/summaryUsecase.go
@@ -1,0 +1,1 @@
+package summary
