+++ internal/domain/user/userHandler.go
@@ -1,0 +1,1 @@
+package user
