+++ internal/domain/user/userUsecase.go
@@ -1,0 +1,1 @@
+package user
